mengubah tipe data pada variabel lat dan lon dari Float ke Double

+++ app/src/main/java/com/example/dicodingstory/data/local/entity/StoryEntity.kt
@@ -21,8 +21,8 @@ data class StoryEntity(
     var createdAt: String? = null,
 
     @ColumnInfo(name = "lat")
-    var lat: Float? = null,
+    var lat: Double? = null,
 
     @ColumnInfo(name = "lon")
-    var lon: Float? = null
+    var lon: Double? = null
 )
